Salt: flatten the directory structure
