+++ macos/Saltfile
@@ -1,3 +1,3 @@
 salt-ssh:
-  config_dir: ./config
+  config_dir: .
   state_verbose: False
