+++ macos/generic-worker.plist.jinja
