+++ macos/pillar/external_data.py
@@ -1,0 +1,15 @@
+import urllib
+import taskcluster
+
+
+def ext_pillar(_minion_id, _pillar, *_args):
+    secrets = taskcluster.Secrets(taskcluster.optionsFromEnvironment())
+    data = secrets.get("project/servo/tc-client/worker/macos/1")["secret"]
+
+    url = "https://raw.githubusercontent.com/servo/saltfs/master/admin/files/ssh/%s.pub"
+    data["ssh_keys"] = [urllib.urlopen(url % name).read() for name in [
+        "jdm",
+        "manishearth",
+        "simonsapin",
+    ]]
+    return data
