--- macos/states/top.sls
@@ -1,3 +1,0 @@
-base:
-  'mac*':
-    - generic-worker
