+++ macos/top.sls
@@ -1,0 +1,3 @@
+base:
+  'mac*':
+    - macos
